+++ libcaf_core/caf/all.hpp
@@ -30,6 +30,7 @@
 #include "caf/config_value.hpp"
 #include "caf/config_value_reader.hpp"
 #include "caf/config_value_writer.hpp"
+#include "caf/console_printer.hpp"
 #include "caf/const_typed_message_view.hpp"
 #include "caf/deep_to_string.hpp"
 #include "caf/defaults.hpp"